fix: use builder pattern for custom implementations
Registering an feature breaks the builder chain. This commit enables to use the builder pattern, for SDK-like features as custom feature implementations.

The custom server deviate from the way it was intended.
Create AL -> register required feature -> build. In short, the builder
pattern.

The intended way was re-introduced and the not required methods are
flagged as deprecated.

This is an backward compatible change.

Signed-off-by: Heiko Schabert <heiko.schabert@siemens.com>